Refactored all files to the mentor's recommendations
