+++ src/main.js
@@ -1,16 +1,16 @@
 import dayjs from 'dayjs';
 import {TOTAL_EVENTS_COUNT} from './mock/consts';
-import {RenderPosition} from './consts/consts';
-import {renderElement} from './utils/utils';
+import {RenderPosition} from './consts';
+import {renderElement} from './utils';
 import {generateEvent} from './mock/event-create';
-import TripInfoView from './view/site-info/site-info';
-import SiteControlsView from './view/site-controls/site-controls';
-import NewEventButtonView from './view/new-event-button';
-import EventSortView from './view/sort';
-import EventsListView from './view/events-list';
-import EventFormView from './view/event-form/event-form';
-import EventView from './view/event/event';
-import EventsEmptyView from './view/events-empty';
+import SiteInfo from './view/site-info';
+import SiteControls from './view/site-controls';
+import NewEventButton from './view/new-event-button';
+import EventSort from './view/event-sort';
+import EventsList from './view/events-list';
+import EventForm from './view/event-form';
+import Event from './view/event';
+import EventsEmpty from './view/events-empty';
 
 const siteHeaderElement = document.querySelector(`.page-header`);
 const siteMainElement = document.querySelector(`.page-main`);
@@ -22,8 +22,8 @@ const events = new Array(TOTAL_EVENTS_COUNT).fill().map(generateEvent);
 const eventsSorted = events.slice().sort((a, b) => (dayjs(a.date).isAfter(dayjs(b.date)) ? 1 : -1));
 
 const renderEvent = (eventsContainer, event) => {
-  const eventComponent = new EventView(event);
-  const eventEditComponent = new EventFormView(event);
+  const eventComponent = new Event(event);
+  const eventEditComponent = new EventForm(event);
 
   const replaceItemToForm = () => {
     eventsContainer.replaceChild(eventEditComponent.getElement(), eventComponent.getElement());
@@ -61,22 +61,22 @@ const renderEvent = (eventsContainer, event) => {
 };
 
 const renderEventsList = (eventsListContainer, tripPoints) => {
-  const eventsList = new EventsListView();
+  const eventsList = new EventsList();
 
   renderElement(eventsListContainer, eventsList.getElement(), RenderPosition.BEFORE_END);
 
   if (!tripPoints.length) {
-    renderElement(eventsList.getElement(), new EventsEmptyView().getElement(), RenderPosition.AFTER_BEGIN);
+    renderElement(eventsList.getElement(), new EventsEmpty().getElement(), RenderPosition.AFTER_BEGIN);
     return;
   }
 
-  renderElement(eventsList.getElement(), new EventSortView().getElement(), RenderPosition.BEFORE_END);
+  renderElement(eventsList.getElement(), new EventSort().getElement(), RenderPosition.BEFORE_END);
 
   tripPoints.forEach((item) => renderEvent(eventsList.getElement(), item));
 };
 
-renderElement(tripMainElement, new TripInfoView(eventsSorted).getElement(), RenderPosition.AFTER_BEGIN);
-renderElement(tripMainElement, new SiteControlsView().getElement(), RenderPosition.BEFORE_END);
-renderElement(tripMainElement, new NewEventButtonView().getElement(), RenderPosition.BEFORE_END);
+renderElement(tripMainElement, new SiteInfo(eventsSorted).getElement(), RenderPosition.AFTER_BEGIN);
+renderElement(tripMainElement, new SiteControls().getElement(), RenderPosition.BEFORE_END);
+renderElement(tripMainElement, new NewEventButton().getElement(), RenderPosition.BEFORE_END);
 
 renderEventsList(tripEventsBoard, eventsSorted);
