+++ src/view/event-sort.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils/utils';
+import {createElement} from '../utils';
 
 const createSortTemplate = () => {
   return `
@@ -31,7 +31,7 @@ const createSortTemplate = () => {
   `;
 };
 
-export default class EventSortView {
+export default class EventSort {
   constructor() {
     this._element = null;
   }
