+++ src/view/site-controls.js
@@ -1,0 +1,60 @@
+import {createElement} from '../utils';
+
+const createMenuTemplate = (isTable = true) => `
+  <nav class="trip-controls__trip-tabs  trip-tabs">
+    <a class="trip-tabs__btn ${(isTable) ? `trip-tabs__btn--active` : ``}" href="#">Table</a>
+    <a class="trip-tabs__btn ${(isTable) ? `` : `trip-tabs__btn--active`}" href="#">Stats</a>
+  </nav>
+`;
+
+const createFiltersTemplate = () => `
+  <form class="trip-filters" action="#" method="get">
+    <div class="trip-filters__filter">
+      <input id="filter-everything" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="everything" checked>
+      <label class="trip-filters__filter-label" for="filter-everything">Everything</label>
+    </div>
+
+    <div class="trip-filters__filter">
+      <input id="filter-future" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="future">
+      <label class="trip-filters__filter-label" for="filter-future">Future</label>
+    </div>
+
+    <div class="trip-filters__filter">
+      <input id="filter-past" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="past">
+      <label class="trip-filters__filter-label" for="filter-past">Past</label>
+    </div>
+
+    <button class="visually-hidden" type="submit">Accept filter</button>
+  </form>
+`;
+
+const createSiteControlsTemplate = () => `
+  <div class="trip-main__trip-controls  trip-controls">
+    <h2 class="visually-hidden">Switch trip view</h2>
+      ${createMenuTemplate()}
+    <h2 class="visually-hidden">Filter events</h2>
+      ${createFiltersTemplate()}
+  </div>
+`;
+
+export default class SiteControls {
+  constructor() {
+    this._element = null;
+  }
+
+  getTemplate() {
+    return createSiteControlsTemplate();
+  }
+
+  getElement() {
+    if (!this._element) {
+      this._element = createElement(this.getTemplate());
+    }
+
+    return this._element;
+  }
+
+  removeElement() {
+    this._element = null;
+  }
+}
